deployment/requirements.txt
numpy==1.26.4
pandas==2.2.2
db-dtypes==1.3.1
tensorflow==2.17.0
google-cloud-storage==2.18.2
google-cloud-bigquery==3.26.0
google-cloud-aiplatform==1.71.1
Flask
gunicorn